Validación de errores Admin/UserController y Admin/BookingController

+++ aplication/app/Http/Controllers/Api/Admin/UserController.php
@@ -11,16 +11,23 @@ class UserController extends Controller
     public function index()
     {
 
-        return response()->json(User::all());
+        return response()->json(User::latest()->get());
     }
 
     public function store(Request $request)
     {
 
-        $input = $request->all();
-        $input['password'] = bcrypt($input['password']);
+        $data = $request->validate([
+            'name' => 'required|string|min:3|max:100',
+            'email' => 'required|email|unique:users,email',
+            'password' => 'required|min:6',
+            'role' => 'required|in:admin,client',
+            'phone' => 'nullable|string|max:20'
+        ]);
 
-        $user = User::create($input);
+        $data['password'] = bcrypt($data['password']);
+
+        $user = User::create($data);
 
         return response()->json($user, 201);
     }
@@ -34,10 +41,11 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         $data = $request->validate([
-            'name' => 'sometimes|string',
+            'name' => 'sometimes|string|min:3|max:100',
             'email' => 'sometimes|email|unique:users,email,' . $user->id,
-            'role' => 'sometimes|in:admin,client',
             'password' => 'nullable|min:6',
+            'role' => 'sometimes|in:admin,client',
+            'phone' => 'nullable|string|max:20'
         ]);
 
         if (!empty($data['password'])) {
@@ -46,18 +54,37 @@ class UserController extends Controller
             unset($data['password']); // ✅ ahora sí existe
         }
 
+        if (empty($data)) {
+            return response()->json([
+                'message' => 'No hay datos para actualizar'
+            ], 422);
+        }
+
         $user->update($data);
 
         return response()->json($user);
     }
 
 
-    public function destroy(User $user){
+    public function destroy(User $user)
+    {
+
+        if ($user->bookings()->exists()) {
+            return response()->json([
+                'message' => 'El usuario tiene reservas activas.'
+            ], 409);
+        }
+
+        if (auth()->id() === $user->id) {
+            return response()->json([
+                'message' => 'No puedes eliminar tu propio usuario.'
+            ], 403);
+        }
 
         $user->delete();
 
         return response()->json([
-            'message' => 'Usuario eliminado correctamente'
+            'message' => 'Usuario eliminado correctamente.'
         ]);
     }
 }
